requirement.txt
langchain
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
##Vector db
chromadb
faiss-cpu
sentence-transformers